harfanglab_app/pyproject.toml
[project]
name = "harfanglab-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.3",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "drf-spectacular>=0.28.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.1",
]
